Server starts, fetches access_token and shuts down
A new process is created and the flask server is served from the
new child process. As soon as the access_token is received, it is stored in a Queue, the
javascript on the page reloads the page to /shutdown which
shuts down the server. When the server shuts down. The child
process joins with the parent process. The access token is now
retrieved from the Queue.