historical_data() function fetches access_token from file if it exists
Instead of everytime verifying the user, this stores the access_token of
the specific account in a temporary file on disk. So whenever, subsequent
requests are made to historical_data() function. It checks whether the file
is present on the disk or not. If the program finds the file. Then it uses
the access_token otherwise the user is re authenticated.

A feature that is still left to implement is to check the "time" field in the
access_token.file to verify that the token that was fetched is new and not stale.